This revision includes the changes for recognizing which chips should be included in the processing based on the 'group' parameter (default being all chips are included).  Modifications were made to acsData, imageObject and processInput to interpret this properly.
This implementation uses the new 'group_member' boolean attribute for each CHIP, which gets set based on the 'group' parameter.  If True, returnAllChips will return that chip, otherwise it will be ignored, for example.

In addition, the drizCR module was revised to use 'returnAllChips' to get the chips it needs for processing, replacing code which had several critical bugs. WJH



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/development/trunk/BigBlackBox@7741 fe389314-cf27-0410-b35b-8c050e845b92