Added support for *non-circular* exclusions and/or *inclusions* regions. Ticket #1055 Users now can call findobj with regular DS9 region files that must be specified in image coordinates. This support adds new dependencies on pyparsing and pyregion.
In order to facilitate (semi-)automated generation of the region files in image coordinates for each input image from DS9 region file(s) specified in sky coordinates, a new task *MapReg* was introduced.

Removed hard-coded restrictions on 'ratio' and 'theta' parameters passed to 'findobj' from 'ndfind'. Added 'ratio' and 'theta' parameters to the TEAL GUI for 'imagefindpars' and 'refimagefindpars'. Ticket #1053 and Ticket #1054

Corrected a bug in computation of kernel size in findobj.py. Ticket #1053.

Corrected a bug in tweakutils.parse_input related to handling of Python lists of strings with wild characters.




git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/branches/fy14dev@27351 fe389314-cf27-0410-b35b-8c050e845b92