INS testing of MapReg has revealed that region files created by the MapReg are not being saved in the directory specified by the outpath parameter (in that directory only the produced catalog file is saved).
Initial implementation of MapReg was working correctly, however, it was using a "patched" (by me) version of pyregion package. Because there was desire on the part of the set-up not to rely on the "patched" version, I switched from the direct call to pyregion.write() to a workaround function. Incomplete file name was passed to this function call. This call needs to be modified to use full file name including directory specified in outpath.

Ticket #1138.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/trunk@31809 fe389314-cf27-0410-b35b-8c050e845b92