Removed prefix randomization from the temporary files created by astrodrizzle and stsci.skypac.skymatch: static mask and skymatch masks now have fixed fixed names.
This change was due to complaints from INS users that adrodrizzle leaves a large number of temporary files that keeps growing after each run. The reason is that users turn 'clean' off because they *do* want to inspect other temporary files, but they do not want these files to grow in number.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/trunk@32128 fe389314-cf27-0410-b35b-8c050e845b92