Main subject of this commit is introduction of support for aligning large mosaics in which some images may not have direct overlap with the reference image - see Ticket #1144.
Also, this commit adds support for writing sharpness and roundness values to the "coordinate" files.

In addition, it solves a few of minor inconsistencies and bugs such as:
1. Tweakreg would crash if no sources are detected in one or more images (now it would skip over these images);
2. Other minor improvements to the code/help.

Note: This commit is a cumulative set of updates made to the drizzlepac/fy14dev branch in the following commits: r32741, r32850, r32950, and r32951.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/trunk@32957 fe389314-cf27-0410-b35b-8c050e845b92