This is identical to r37066 made to the trunk:
This commit addresses the issues described in r36738. It appears that even the increase of the `maxerr` from 10 to 50 in r36738 was not enough to solve the issue described in r36738. It appears that some HST images that are supposedly distortion-corrected (e.g., "drizzled") have non-orthogonal CD matrices.

In this revision I remove the check that reference images are distortion-free (this was an assertion mostly for debugging purposes). TweakReg will proceed working with reference images that have slight distortions but the result may be less than perfect.

This revision also simplifies the check for matrix orthogonality.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/branches/irafx@37067 fe389314-cf27-0410-b35b-8c050e845b92