DrizzlePac v2.1.0 (Major Update)
This update includes revisions to address the following tickets:

#1182 - Both astrodrizzle and tweakreg now return an output CD matrix which has 
identical cross-terms indicating the same scale and orientation in each axis. 
This relies on a revision to stwcs.distortion.utils.output_wcs() found in r45804.

#1169 - User-provided list of input catalogs no longer needs to be matched 
exactly with input files. As long as all input images are included in input 
catalog list in any order, tweakreg will apply the correct catalog to the 
correct file.

#1231 - The user interfaces to all 3 coordinate transformation tasks now use 
'coordfile' as the input file of coordinates to transform. The use of 'coords'
has been deprecated, but still can be used if needed.  However, use of 'coordfile' 
will always override any input provided simultaneously with 'coords' parameter. 
Help files have been updated to document this as clearly as possible for users.

#1135 - Help files have been updated to document API changes from #1231 and 
#1217. 

#1147 - Tweakreg has been updated to correctly and fully apply source selection
criteria for both input source catalogs and reference source catalogs based
on fluxmin,fluxmax and nbright for each. 

#1232 - All use of keyword deletion has been updated in drizzlepac (and 
fitsblender) to avoid warnings from astropy.

#1217 - Use of 'updatewcs' through the Python interface for both the 
astrodrizzle.Astrodrizzle() and tweakreg.TweakReg() functions was fixed.

#1233 - All 3 coordinate transformation tasks rely on the input of valid WCS
information for the calculations.  These tasks now warn the user when it could
not find a valid WCS and instead defaulted to using a unity WCS, so that the user
can understand what input needs to be checked/revised to get the correct results.

-- The reported fit as written out to a file has been slightly modified
to report more appropriate numbers of significant digits for the results.

-- Use of astrolib.coords was removed from drizzlepac and replaced by use of 
astropy functions instead. This eliminated one more obsolete dependency in our
software.



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/drizzlepac/trunk@45805 fe389314-cf27-0410-b35b-8c050e845b92