Modified the handling of the filter values to accommodate SBC names (#586)
* Modified the handling of the filter values to accommodate SBC names
which are currently malformed (e.g., "F140LP;N/A;F140LP". The third
token delimited by ";" should not be present. Also, the "N/A" needs
to be replaced with "none" as the slash cannot be part of the output filename.

* Fixed line indentation so routine works properly for all cases.

* Fix filename creatation from FITS keyword values and defined rules.
Also, if the only viable FILTER keyword value is "BLOCK*", then the
dataset cannot be aligned.

* Removed unnecessary print statement.