Created a module, svm_quality_analysis.py, to encapulate the functions (#625)
which compute a variety of statistics on the SVM data products. The
evaluation output is saved to JSON files. This checkin only generates
the initial function, compare_num_sources(). The functions in this module
are intented to be called from pipeline software, or from the command line
after the SVM products have been produced. A driver to this module is
to follow in another PR.