Update to Segmentation SVM processing via use of RickerWavelet kernel for crowded fields (#756)
* Fixed a bug in hapsequencer module when only processing for the segmentation
catalog such that the "phot_mode" is only "segment".
Added/modified a few log.info entries in astrometric_utils module to clarify
detection kernel used.
catalog_utils module:
Added code from R.L.White which evaluates the segmentation image and determines
if the field is crowded.  Reorganized code in identify_sources method as there
may be an iteration to detect sources with an alternative kernel for crowded fields.
Created the detect_and_deblend method to isolate the detecting/deblending.
If the evaluation of the segmentation map indicates crowding, a RickerWavelet2DKernel
is defined, and an alternative segmentation map is generated.

* catalog_utils.py
- Track the invocation of the detect_and_deblend() method. In order to
  generate "numbered" segmentation maps when in diagnostic mode.
- When evaluating the segmention map generated by using the
  RickerWavelet kernel, modify the "big_island" and "mask_biggest_source"
  parameters.
- Deep copy the final segmentation to be used to a class attribute.
- Ensure the proper segmentation map, kernel, and fwhm values are used
  appropriately.
- Experimental: Modify the threshold value in use from the nominal
  algorithm by adjusting the nsigma value.

* catalog_utils.py module
- Added commentary.
- Second call to segmentation image evaluation is to find size of largest
  "island".
- Allow the source box size to be a parameter and not the class attribute
  when calling detect_and_deblend().
- Do not normalize the RickerWavelet kernel.

* - Added variables in support of the RickerWavelet kernel to the
  instrument/detector configuration files.
- Read the RickerWavelet values from the configuration files.
- Add warning to log and output Segmentation catalog if the
  RickerWavelet segmentation image fails the "big island" test,
  since the image is still used for processing.

* - Fixed an oversight in hapsequencer module to ensure "point"
catalog-specific code is not invoked when only processing for
"segment".

* Remove unnecessary statements regarding default values for parameters
as the values can be seen in the method signature.