Re-organize/streamline background computation algorithms (#827)
* Initial re-work of the compute_background method of the CatalogImage
class.  WORK IN PROGRESS ONLY.

* Next iteration of update to compute_background in CatalogImage class.
Removed deprecated code from HAPSegmentCatalog with respect to the
background computations.

* Third iteration of background fixup - still a WIP.  Code was run
on a single example dataset, ibe809.  Configuration values are
still hard-coded or in the catalog-specific portions of the
parameter files.  Validation of the algorithm still needs to
be done.

* Added code to re-compute the threshold above which sources are
detected when calling the RickerWavelet algorithm for the
Segmentation catalog.

Updated how the configuration parameters are accessed (general vs
algorithm-specific), and correspondingly, updated the configuration
files to reflect the same change.

Set the "scale" configuration value for ACS/SBC and WFC3/IR to 100.0
so there is essentially only one weighted region.