Fix a bug introduced when cleaning up two-dimensional images (#828)
* Changed or added where "del" is used in the compute_background
image method for cleanup.  One of the del commands was also a
bug when used in the mode of "simple_bkg=True".  Deleted some
redundant code.

* Fixed bug to ensure when the "simple_bkg = True", only the sigma-clipped
background is computed.  Updated "del" when two-dimensional images can
be deleted, and updated the "negative_threshold" from -0.5 to 0.0 in the
determination of negative data values in the background-subtracted image.