Add computation for the minimum allowable threshold  (#837)
* Added the computation for the minimum allowable threshold based upon the
readnoise, gain, number of images, and total exposure time.  This minimum only
applies for the sigma_clipped_stats algorithm where a comparison is done between
the rms computed by the sigma_clipped_stats and the minimum allowable threshold.
Clean up with respect to images written in diagnostic mode.  Fixed a bug where
a residual computation of the threshold had not been deleted.

* Segmentation algorithm only: Added evaluation of segmentation image
before deblending is done. This is only informational at this time.
Fixed the weighting for the computation of the threshold.

* Change name of class attribute from self.exclusion_mask to self.exclude_zones_mask
so there is no confusion between the attributes of CatalogImage and HAPCatalogBase.
Change docstring from Returns to Attributes.