Address identically zero background seen in SBC images (#846)
* Pass the total object project "footprint" mask attribute to the HAPCatalogs class.
Streamline computations in the compute_background() method of CatalogImage, in particular
use boolean footprint and inverse footprint masks for efficient computation.  Add code
to accommodate images with identically zero backgrounds (e.g., some SBC images) and modify
other portions of the code to handle this particular case.

* - Removed temporary debug code (e.g., output of FITS images).
- Made a threshold (limit on the percent of zeros) a configuration variable.
- Removed all references to exclusion_mask or exclude_zones_mask and replaced
  with footprint_mask or inv_footprint_mask.
- Streamlined some computations.

* Updated one call to compute_background() which did not have the
zero_percent argument.

* New parameter for the build_kernel() method.

* Re-instate the original first parameter for creating the weight masks as
the attribute of the CatalogImage class and based on the WHT extension of
the input data. the local variable wht_image is unnecessary.

* Make sure the image to the HAPPointCatalog has no nans.