Fixed two bugs during SVM testing: attribute not part of a class and error in joining tables (#898)
* Fixed test for "big island" message as it only applies for the segmentation algorithm.
Modified join of specified filter columns to be added to the total catalog - now a "left"
join.  Modified construction of Point total catalog with respect to the RA and Dec columns
to ensure these columns are always present and totally populated.  This change was necessitated
by the change to the type of join now being performed.

* Fixed bug.  Point catalog column name could actually one of two
possible values.