Address problem of the deblend step taking too long (#917)
* - Re-organized the code so detection, deblending, and evaluation of the segmentation
images are all now independent methods for purposes of flexibility.
- Algorithm now detects the segments, but it does not immediately deblend the
image.  The segmentation image is now evaluated for big islands/crowded fields (bi/cf).
If the segmentation image is determined to have bi/cf, then the RickerWavelet (RW) alternative
is immediately invoked.  This was done based upon examination of the "1000" test
datasets and deblending times.  Only after the RW segmentation image is generated, the
deblending is performed.
- Renamed all RickerWavelet configuration variables to have a rw2d prefix
and modified the associated code.

* Separate evaluation of segmentation image into its own method.
Fixed typo of parameter to deblend_segments().

* Clean-up and streamline of previous updates.

* Clean up of formatting.