Handle situation of no viable Total Data Products in the visit (#946)
* Handle situation of no viable Total Data Products in the visit (e.g., only Grism data and no direct images).

* Handle the case where the visit consists of only Grism exposures - no processing
done, ensure there is an output trailer file, remove the SVM FLT/FLC images from
disk, and the manifest contains only the trailer file.  Added log messages as
appropriate.  Added checks in some routines in hapsequencer requiring indentation
for an "if" statements. Fixed deletion of Ramp exposure objects from Product lists
and ensure the Ramp images are in the manifest.

Co-authored-by: Michele De La Pena <mdelapena@hlalab1.stsci.edu>