Improved the way WCS solutions are manipulated for visits with Grism/Prism data (#956)
* Modified the way WCS solutions are manipulated.  In particular, when the
active WCS solution is being replaced, the current active solution (and
any alternative WCS solutions from the same distortion model) are saved
as headerlet extensions (if they are not duplicates according to their
HDRNAME).

* Ensure a trailer file is written for the Grism/Prism and Ramp images.
The trailer files will be copies of the total data product trailer file.

* Implemented some suggestions from review.  More thorough testing found a
bug in the checking of alternate WCS solutions.  Implemented a more Pythonic
way to perform some actions, and a bit of cleanup with code and messages.