Debugging issues with archiving and restoring WCS solutions from headerlets (#963)
* Debugging issues with archiving and restoring WCS solutions from
headerlets.

* (1) Addressed restoration of WCS from headerlet when there are duplicate
headerlets (as indicated by same HDRNAME and WCSNAME) for some Grism/Prism
data.  One of the duplicate headerlets is deleted.  (2) Trap exception
when cannot restore a WCS from headerlet as the HDRNAME keyword cannot
be found.  The exception is reported, and a "warning" message is used
that the WCS cannot be restored for the Grism/Prism data.  However,
processing can and should continue for the remainder of the Grism/Prism
and direct images.