WIP: Modified align_to_gaia() method of the FilterProduct to use logic provided by Astrometry WG (#1002)
* WIP: Modified align_to_gaia() method of the FilterProduct. Automatically
tries to perform matching and fit using a method list and catalog list.
The first satisfactory fit is used.  Need the evaluation criteria for
a "successful" fit, and possibly put some values into a configuration
file.

* WIP: Modified code based upon code review: make the reference catalog name
catalog-specific so issues can be traced.  Also, added configurable
variables to the JSON configuration files.

* Incorporated the use of the new functionality which determines the fit
quality.  Ensured the proper configuration parameters are used for the
fitting, and updated the catalogs to use for SVM/MVM processing.