WIP: Address unreasonable deblending times for segments  (#1067)
* WIP: Restructed code to allow for modest recomputation with regard to identifying
segments with the custom/Gaussian or RickerWavelet kernel.  If the segmentation
image is found to have big islands or a large source fraction after both
kernels have been tried, the algorithm allows for a recomputation of the background
image (sigma-clipped to Background2D) or raising the threshold for detection if the
Background2D is already in use.  Pending: last ditch thresholds on big islands and/or
source fraction which allow segmentation processing to continue as the deblending
should statistically not take days and days (based upon test results).  I still
need to analyze the stats to determine these thresholds.

Reset some configuration variables, keep track of the background determination
type (zero_background, sigma_clipped_background, twod_background), reset class
attribute to a local variable, switch order of catalog computation (segmentation is
now done first), ensure use of the "custom kernel" for the segmentation detection.
Currently an incomplete update of all configuration files.

* Added code to either switch to using the Background2D to determine the background
or increase the threshold level for source detection if already using Background2D.
Also, added "last effort" to keep going using new limits on the big sources/islands
and source fraction.  These larger limits allow some big islands or a larger
source fraction based upon statistical testing the deblending of such data will
take a reasonable amount of time (< 1 day).  These limits are still to be
verified.  Updated all ACS and WFC3 configuration files for the new "deblend
limits" and increased the rw2d_sigma (threshold detection).

* Added recomputation of the Custom/Gauassian and the RickerWavelet kernels
when a new background type is computed.  Added some necessary log statements
and made them Warnings.

* Based on discussion with Warren, created the ratio of big sources found
when using Custom/Gaussian kernel versus the Rickerwavelent kernel.  When
this ratio is > 2, this seems to be indicative of overlapping PSFs in the
image versus large areas of nebulousity. The overlapping PSFs may occupy a
large source fraction of the image, but the segments deblend quickly. The
recent improvements involve balancing the desire to produce a viable
segmentation catalog vs deblending segment(s) which can take a VERY long
time and not produce a good result in the end.