Fixed bug to access the dataset filenames properly AND remove zero exposure images from processing (#1081)
* Fixed the bug to access the dataset filenames properly if all Grism/Prism
datasets need to be deleted from the processing.  Also, fixed an unreported
algorithmic bug causing the ibbr01 to not produce output.  Two of the
Grism/Prism data files actually had an exposure time of zero.  The
Grism/Prism data is minimally processed in order to have the same
WCS as the corresponding direct images. STWCS found the images had
zero exposure time, so the images had only an OPUS WCS. This
caused the processing to fail as there was no preferred 'a priori'
WCS available for all the images to use.

* Fixed typo in list name.