Added functionality to combine image cutouts, mvm_combine. This … (#1110)
* WIP: Added functionality to combine image cutouts, mvm_combine. This new function required
modifying the make_the_cut routine to retain some keywords written by the low-level
infrastructure software, astrocut.  Clean up some output files written by astrocut which
we do not want to be retained.

* Cleaned up cruft to move source out of WIP.  Allowance for a custom
combine routine will be part of a new ticket.

* Addressed comments in review.  Also, ensure in the situation of multiple cutout
output files, the primary header in the input is the primary header in the
output file, except for a few extra keywords.