SVM: Handle multiple total data products (TDP) when first TDP processed has no sources (#1115)
* Added informational log messages regarding lack of sources found and
catalog status. Protect against divide by zero when computing the
ratio of "Gaussian big islands" vs "RW big islands".  Ensure the
segmentation image is not None upon return from detect_and_eval_segments()
and handle the situation appropriately.

* Change "else" to "elif".

* Initialize variable.

* Set the default values for crowding, biggest source, and source fraction so
they are consistent throughout the module.  Fixed the name of a variable and
set the default for the variable to allow for deblending.  The variable is a
fraction and needed to accommodate the unlikely situation where the denominator
is identically 0.0.