Function mvm_analyze_wrapper determines the viability of an image for MVM processing (#1132)
* Added a new function, mvm_analyze_wrapper, which uses the functionality in the underlying
analyze_data routine to examine FITS keywords and determine the viability of the data to
be used in the MVM processing.  The mvm_analyze_wrapper works on a single image at a time.
Added an optional parameter to analyze_data so the routine can determine when MVM processing
is wanted as MVM processing does not allow anything to be done for GRISM/PRISM images.
Updated doc strings.

* Changed logging level from DEBUG to NOTSET.

* Changed log setting to DEBUG in the creation of the logger, as well as setting the
level for the handlers to force out INFO messages. Improved documentation of an input
parameter.  Made logic more explict in the code handling Grism/Prism data.