Add the Kron radius as an output measurement for the catalog (#1142)
* Added the Kron radius as an output measurement for the catalog.  This change required providing
"kron_params" to the SourceCatalog instantiation.  The "kron_params" are a list of two values that
are used to determne how the Kron radius and flux are calculated.  The two values, kron_scaling_radius
and kron_minimum_radius, are stored as two new configuration values in the catalog configuration file
associated with each of the detectors.

* Need to commit the updated detector catalog configuration files to accompany the Kron radius
computation.

Co-authored-by: Michele De La Pena <mdelapena@dlhlalab2.stsci.edu>