Update release 3.3.x for release candidate 3  (#1159)
* Fixed a bug which caused the proper naming of the output manifest (#1146)

file to be botched.  The manifest filename is now properly
constructed if the main input is a poller file (no
specific filename format expected), a user-defined filename
containing a list of files, or a Python list.  In the latter
two cases, the manifest filename is constructed by interrogating
the primary header of the first file for information.  In the
poller file case, the total data product has a manifest
filename attribute which is used.  If something goes completely
amiss, the manifest filename will be "manifest.txt".  In all other
cases the manifest filename will be of the form
instrument_programID_obsetID_manifest.txt (e.g.,wfc3_b46_06_manifest.txt).

Co-authored-by: Michele De La Pena <mdelapena@dlhlalab2.stsci.edu>

* Changed names of "ISO" columns in Segmentation catalog to be unique in (#1155)

order to avoid confusion with users familiar with SourceExtractor and its
variables (MagIso->MagSegment, FluxIso->FluxSegment, and FluxIsoErr->FluxSegmentErr).

* Fix use of filename in get_sky_cells (#1158)

Co-authored-by: mdlpstsci <mdelapena@stsci.edu>
Co-authored-by: Michele De La Pena <mdelapena@dlhlalab2.stsci.edu>