Initial draft of the descriptive documentation MVM hapcut utilities (#1150)
* Initial draft. Created descriptive documentation for the
hapcut_utils.py module.  Described at a high-level: how to use
the utilities, actual functionality of the utilities, input and
output filename nomenclature, and contents of output files.

* Many changes to the documentation, as well as to the corresponding code (done
in a different branch).  The changes were in response to the initial code review.
Further refinement is needed for the full worked example, particularly with
respect to the display of the images.