Improvements/updates made to MVM utilities while documenting the code (#1148)
* For mvm_id_filenames, make the output MAST query results file have a
consistent naming convention with the remaining hapcut utilities.
For make_the_cut, loop over the files for which cutouts have been requested.
In this way if one of the files causes an exception for any reason (e.g.,
the file does not contain the cutout region), an error is issued, but
the remaining files continue to process. For make_the_cut and mvm_combine,
added the instrument designation to the output files for completeness (though
now the filenames are even longer).  For mvm_combine, added the FILENAME
keyword to the PHDU of the output combined file.  Also, improved the EXTNAMEs
of all the EHDUs to reflect the data they contain.  Added or improved clarity
of comments.

* Changes to fix bugs, force the output filenames to be consistent
with other output MVM products, abbreviate the complex EXTNAMEs in
the cutout and combo files, and update documentation so it is
clear the pipeline does not produce exposure-level MVM drizzled
products, but the user can do this with a custom processing.

* Modified the make_the_cut function to read files not in the current working directory and
to utilize the output_dir parameter to write files in the specified directory.