Added "demonstration" routine, test_svm_demo.py, which uses fixtures (#1185)
* Added "demonstration" routine, test_svm_demo.py, which uses fixtures
for basic functionality.  This is a WIP.  The poller file needs to
be in the directory with the tests.

* Modified to use IPPPSS plus a wildcat to request data for download from
MAST.  This is to ensure individual files which are part of an ASN can
be accessed.

* Make sure to give the full path to the input data file (poller file)
everywhere it is referenced in the test module.  The file lives with
the tests.

* Replaced the dataset in use so the desired wcsname contains "FIT_SVM_GAIA".

* Updated test_svm_demo.py was not part of the previous commit which
was a mistake.  Here it is now!

* Corrected the source code to do as intended and only check the *total*.fits
files for wcsname.  Renamed test_svm_demo.py to template_svm_demo.py as it
is intended to be a demonstration file and not actually run as a test.  Created
test_svm_je281u.py which is the actual data test.

* Added "force=True" to the logging.basicConfig to force shutdown of logging
handlers when running from PyTest.  Updated comments in the template and
the je281u test.

* Delete acs_8f6_61_input.out

This poller file should not have been part of the PR in the first place.