Added a new PyTest test_svm_j97e06  (#1189)
* Added a new test test_svm_j97e06 which is ACS WFC, Grism and Direct image data.

* Add the poller file acs_97e_06_input.out to go with its PyTest.

* Modified the read of the manifest file from ascii.read to Table.read
(both with format="no_header") in an effort to overcome the PyTest objection.

* Change made only as a debug test of the Azure test system.

* Another test to debug failure under Azure.

* Cleared debug statements and use alternate method to read the manifest
file (with and readlines).  It is not understood why the Table.read or
ascii.read did not work.

* Changes to test_svm_j97e06.py to work with --basetemp and added comments.
Also changed test_svm_je281u.py and template_svm_test.py for the --basetemp
to make all the files consistent (a but of smuggling).

* Review suggests to take care to strip the end of a string of characters
when later using the string with "endswith".

* Per review - Modified to not perform an assert in a loop as this would
pre-maturely end the processing.  Fixed failure messages.