Reworked the analyze_wrapper function to accommodate whether or not to use ACS SBC or HRC data for MVM processing based upon environment variables.  Code which did this previously was removed from the poller_utils module.  The update to analyze_wrapper naturally allowed for use of return codes other than 0 or 1 for Operations to distinguish why datasets did not process (55: SBC/HRC and other non-viable data, 65: non-viable data).
Additional log messages were added indicating why datasets would not be processed,
as well as how many datasets were found that would be processed.

As sys.exit() is used to cease processing, an additional "except" branch of
"except BaseException" was added to hapmultisequencer.py to catch the return
codes and handle generating empty manifest files with proper names to meet
expectations of Operations.