Special handling of SBC/HRC MVM data, clarify messsages, return codes (#1211)
* Reworked the analyze_wrapper function to accept two new parameters, an
indicator for SVM or MVM processing and an indicator as to whether or
not SBC/HRC data should be used for MVM processing.

The MVM indicator has two uses: (1) Marks Grism/Prism data as "do not
process" in the filtered table associated with the poller file, and
(2) acts as a gate for further checking of the filtered table data when
SBC/HRC data should not be used for processing.

Clarified messages which are generated for the log for the cases with
associated return codes:

 No viable images in single/multi-visit table - no processing done.
 Return code = 65

 Only non-viable or SBC/HRC images in the multi-visit table - no processing done.
 Return code = 55

* Reworked the analyze_wrapper function to accommodate whether or not to use
ACS SBC or HRC data for MVM processing based upon environment variables.  Code
which did this previously was removed from the poller_utils module.  The update to
analyze_wrapper naturally allowed for use of return codes other than 0 or 1 for
Operations to distinguish why datasets did not process (55: SBC/HRC and other
non-viable data, 65: non-viable data).

Additional log messages were added indicating why datasets would not be processed,
as well as how many datasets were found that would be processed.

As sys.exit() is used to cease processing, an additional "except" branch of
"except BaseException" was added to hapmultisequencer.py to catch the return
codes and handle generating empty manifest files with proper names to meet
expectations of Operations.