Address issue of logging for PyTest tests (#1247)
* Added better traceback information to the Exception in product.py.  Added/renamed
a pytest: test_svm_wcsonly.py to test_svm_canary.py and renamed its associated input
file containing a list of filenames: test_acs_hrc_sbc_input.out to acs_hrc_sbc_input.out.

* Modified the align.py module, perform_align() and determine_fit_quality(), by defining global
variables to use for the log file handler and log filename.  This was done to address the PyTests
from having an error looking for a directory associated with the first test as all PyTests are
done in a single Python session.  The error would cause the test to fail, and the tests need
to be executed to ensure the algorithms continue to work as expected.  Essentially, the log
file handler is reset as necessary so a fresh handle is assured for each new test.  Some PyTests
were updated to PASS.  Updating the tests further is a topic for a different PR.

* Changes done as a result of PR review: (1) PyTest updates to make some comparisons
more efficient by using NumPy.isclose, and (2) try/except the os.remove() which
could fail on Windows.  There is no impact here for Operations as this would be
on a Windows system.