Improve algorithm for identifying and filtering large segments (#1357)
* Improved the algorithm used to filter the array of big segments which
will be deblended.  Added safety catches, more comments, and clean-up to
the logic.  The clean-up clarifies the distinction between the code
used to identify "big islands" or a large "source fraction" which are
only used to set a flag to indicate another kernel or background
computation could yield improved results.  This is distinct from
an array of identified large segments which are filtered only to
contain segments of reasonable size to be deblended.  Deblending
enormous segments can be prohibitive in time.

* Further code to ensure computations in "for" loop properly take into
account the entries in the array being processed since the first entry
has been handled independently. The internals of the "for" loop look
ahead.