Only allow "verify_guiding" check for MVM processing (#1368)
* Added protection around the function call, verify_guiding, from the
analyze_wrapper routine.  The function will only be called if the
processing type is "MVM".  Also added a new parameter, do_verify_guiding,
to the process function in runastrodriz.py.  This new parameter will
allow a user to turn on verify_guiding for pipeline processing.  Organized
the order of the parameter switches for the process function.  Added
explanatory comments in the code for these changes.

* Fix a bug so the "-a" option sets "force_alignment=True".