Generate 342rc1 (#1373)
* This change addresses bugs associated with the big_segments attibute of (#1365)

the segmentation image. (1) During construction of the attribute, it is
necessary to check the array is not empty.  If it is, set the attribute
to None.  (2) Only do the deblending step if the attribute in not None.

* Update the WFC3 rules files to remove WCS-associated keywords (CRPIX, CRVAL, (#1366)

CTYPE, and CD) from the three column entries which govern how the output header
values are computed.

* Ensure the HAPEXPNAME column is the size of the largest input name (#1371)

currently being processed OR at least 51 characters.

Co-authored-by: mdlpstsci <mdelapena@stsci.edu>