Modified the "exposure level" products (ExposureProduct, GrismExposureProduct, (#1398)
and SkyCellProduct) to ensure the products have a HAPLEVEL = 1.  The copy
method of each of these classes set the HAPLEVEL to 0 in anticipation
this would be updated later in the code.  No update was ever done, so
the HAPLEVEL has been incorrect for the SVM exposure level FLT/FLC files.