Created the mvm_quality_analysis.py module  (#1396)
* Created the mvm_quality_analyis.py module which generates JSON
information regarding the MVM processing.  This module is a streamlined
version of the svm_quality_analysis.py module in that it has been
cleaned up to contain only what is necessary to run the initial
data collection routine.

* Only process the "fine" scale WFC3 IR data products.  Create a dictionary of
WCSNAMEs so each exposure can be assigned a numeric value of 2^n (n=0,1,2...)
which is statically associated with the WCSNAME. The "better" WCSNAMEs have the
lower 2^n values.  Retain the running image values which are also 2^n as these
values can be used to identify uniquely each exposure (both chip).

* Changes made to the descriptions of JSON variables and added a comment
regarding future use of WFPC2 data.